DRTII-924 Refactor selectable ports by region into component for reuse  - Use for signup form  - Use for alerts form
