+++ react/src/components/Alerts/Alerts.tsx
@@ -5,10 +5,12 @@ import {a11yProps, TabPanel} from "../TabPanel";
 import AlertForm from "./AlertForm";
 import {ListAlerts} from "./ViewAlerts";
 import {UserProfile} from "../../model/User";
+import {PortRegion} from "../../model/Config";
 
 moment.locale("en-gb");
 
 interface IProps {
+  regions: PortRegion[]
   user: UserProfile;
 }
 
@@ -24,7 +26,7 @@ export default function Alerts(props: IProps) {
       <Tab label="View Alerts" {...a11yProps(1)} />
     </Tabs>
     <TabPanel value={selectedTab} index={0}>
-      <AlertForm user={props.user} callback={() => setSelectedTab(1)}/>
+      <AlertForm regions={props.regions} user={props.user} callback={() => setSelectedTab(1)}/>
     </TabPanel>
     <TabPanel index={1} value={selectedTab}>
       <ListAlerts/>
